refactor: move the custom hooks to a separate files
