+++ src/navigation/components/Navigation/Navigation.tsx
@@ -1,17 +1,10 @@
-import React, {useEffect, useMemo, useState} from 'react';
-
-import debounce from 'lodash/debounce';
+import React, {useState} from 'react';
 
 import OutsideClick from '../../../components/OutsideClick/OutsideClick';
 import {Col, Grid, Row} from '../../../grid';
-import {
-    ClassNameProps,
-    HeaderData,
-    NavigationItemModel,
-    ThemedNavigationLogoData,
-} from '../../../models';
+import {useActiveNavItem, useShowBorder} from '../../../hooks';
+import {ClassNameProps, HeaderData, ThemedNavigationLogoData} from '../../../models';
 import {block} from '../../../utils';
-import {getNavigationItemWithIconSize} from '../../utils';
 import DesktopNavigation from '../DesktopNavigation/DesktopNavigation';
 import MobileNavigation from '../MobileNavigation/MobileNavigation';
 
@@ -23,51 +16,6 @@ export interface NavigationComponentProps extends ClassNameProps {
     logo: ThemedNavigationLogoData;
     data: HeaderData;
 }
-
-export const useActiveNavItem = (
-    iconSize: number,
-    leftItems: NavigationItemModel[],
-    rightItems?: NavigationItemModel[],
-) => {
-    const [activeItemId, setActiveItemId] = useState<string | undefined>(undefined);
-
-    const getNavigationItem = getNavigationItemWithIconSize(iconSize);
-
-    const leftItemsWithIconSize = useMemo(
-        () => leftItems.map(getNavigationItem),
-        [getNavigationItem, leftItems],
-    );
-    const rightItemsWithIconSize = useMemo(
-        () => rightItems?.map(getNavigationItem),
-        [getNavigationItem, rightItems],
-    );
-
-    const onActiveItemChange = (id?: string) => {
-        setActiveItemId(id);
-    };
-    return {activeItemId, leftItemsWithIconSize, rightItemsWithIconSize, onActiveItemChange};
-};
-
-export const useShowBorder = (withBorder: boolean, withBorderOnScroll: boolean) => {
-    const [showBorder, setShowBorder] = useState(withBorder);
-
-    useEffect(() => {
-        if (!withBorderOnScroll) return () => {};
-
-        const showBorderOnScroll = () => {
-            if (!withBorder) {
-                setShowBorder(window.scrollY > 0);
-            }
-        };
-
-        const scrollHandler = debounce(showBorderOnScroll, 20);
-
-        window.addEventListener('scroll', scrollHandler, {passive: true});
-        return () => window.removeEventListener('scroll', scrollHandler);
-    });
-
-    return [showBorder];
-};
 
 export const Navigation: React.FC<NavigationComponentProps> = ({data, logo, className}) => {
     const {
